imageupload

+++ database/migrations/2024_10_03_114624_create_posts_table.php
@@ -13,16 +13,20 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('posts', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreignId('category_id')->references('id')->on('categories')->onDelete('cascade');
-            $table->string('title');
-            $table->longText('description')->nullable();
-            $table->boolean('status')->default(false);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('posts')) {
+            Schema::create('posts', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('gallery_id')->references('id')->on('posts');
+                $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreignId('category_id')->references('id')->on('categories')->onDelete('cascade');
+                $table->string('title');
+                $table->longText('description')->nullable();
+                $table->boolean('status')->default(false);
+                $table->timestamps();
+            });
+        }
     }
+
 
 
     /**
